VaultDetailBalanceContent refactored
