+++ VultisigApp/VultisigApp/Views/Components/VaultDetailBalanceContent.swift
@@ -16,8 +16,15 @@ struct VaultDetailBalanceContent: View {
     @EnvironmentObject var homeViewModel: HomeViewModel
     
     var body: some View {
+        container
+            .onAppear {
+                setData()
+            }
+    }
+    
+    var content: some View {
         ZStack {
-            content
+            balanceContent
             hideButton
         }
         .frame(maxWidth: .infinity)
@@ -25,15 +32,9 @@ struct VaultDetailBalanceContent: View {
         .background(Color.backgroundBlue)
         .listRowInsets(EdgeInsets())
         .listRowSeparator(.hidden)
-#if os(macOS)
-        .padding(.vertical, 18)
-#endif
-        .onAppear {
-            setData()
-        }
     }
     
-    var content: some View {
+    var balanceContent: some View {
         Text(homeViewModel.hideVaultBalance ? redactedText : (homeViewModel.selectedVault?.coins.totalBalanceInFiatString ?? ""))
             .font(.title32MenloBold)
             .foregroundColor(.neutral0)
@@ -42,44 +43,6 @@ struct VaultDetailBalanceContent: View {
             .frame(width: width)
             .lineLimit(1)
             .minimumScaleFactor(0.5)
-    }
-    
-    var hideButton: some View {
-        VStack {
-            Button {
-                withAnimation {
-                    homeViewModel.hideVaultBalance.toggle()
-                }
-            } label: {
-                Label("", systemImage: homeViewModel.hideVaultBalance ? "eye.slash" : "eye")
-                    .labelsHidden()
-                    .foregroundColor(.neutral0)
-                    .font(.body16Menlo)
-            }
-            .contentTransition(.symbolEffect(.replace))
-        }
-        .font(.largeTitle)
-#if os(iOS)
-        .offset(x: width/2 + 28 , y: 3)
-#elseif os(macOS)
-        .offset(x: width/2 + 16 , y: 3)
-#endif
-    }
-    
-    private func setData() {
-        let balance = vault.coins.totalBalanceInFiatString
-#if os(iOS)
-        width = balance.widthOfString(usingFont: UIFont.preferredFont(forTextStyle: .extraLargeTitle))
-#elseif os(macOS)
-        width = balance.widthOfString(usingFont: NSFont.preferredFont(forTextStyle: .title1))*2
-#endif
-        
-        redactedText = "$"
-        let multiplier = Int(width/25)
-        
-        for _ in 0..<multiplier {
-            redactedText += "*"
-        }
     }
 }
